tests/requirements.txt
coverage>=3.7.1
aldryn-apphook-reload
djangocms-helper
django-parler>=1.9.2
flake8
python-coveralls>=2.9.1
six
tox
python-slugify
